Move existence check to resourcefactory
